# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/episodes": {
      "post": {
        "operationId": "createEpisode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEpisodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR`: body mal formado (falta un campo, tipo incorrecto o campos de más).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "`VOICE_NOT_CONFIGURED`: falta la voz de algún agente candidato para el idioma pedido y el proveedor de TTS activo. No se crea nada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      },
      "get": {
        "operationId": "listEpisodes",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtro por estado: valores de EpisodeStatus separados por coma (por ejemplo `PENDING_REVIEW,REQUIRES_HUMAN_REVIEW`). Un valor que no es un EpisodeStatus es 400 `VALIDATION_ERROR`.",
            "schema": {
              "pattern": "^(?:\\s*(?:CREATED|RESEARCHING|READY_FOR_DEBATE|DEBATING|JUDGING|PENDING_REVIEW|APPROVED|CANCELLED|FAILED|GENERATING_AUDIO|READY_FOR_RENDER|RENDERING|COMPLETED|REQUIRES_HUMAN_REVIEW)\\s*(?:,\\s*(?:CREATED|RESEARCHING|READY_FOR_DEBATE|DEBATING|JUDGING|PENDING_REVIEW|APPROVED|CANCELLED|FAILED|GENERATING_AUDIO|READY_FOR_RENDER|RENDERING|COMPLETED|REQUIRES_HUMAN_REVIEW)\\s*)*)?$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpisodeListItemDto_Output"
                  }
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR`: algún valor del filtro `status` no es un EpisodeStatus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/episodes/{id}": {
      "get": {
        "operationId": "getEpisodeDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetailDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: el episodio no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/episodes/{id}/audio/{audioAssetId}/url": {
      "get": {
        "operationId": "getEpisodeAudioUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "audioAssetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "URL firmada de `/audio-files`, vence a los `AUDIO_URL_TTL_SECONDS`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedAudioUrlDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: el episodio no existe o el `audioAssetId` no es de un argumento de este episodio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/episodes/{id}/manifest": {
      "get": {
        "operationId": "getEpisodeManifest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemotionManifestDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: el episodio no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "`MANIFEST_NOT_READY`: todavía no hay audio o veredicto para todos los argumentos oficiales.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/episodes/{id}/actions/approve": {
      "post": {
        "operationId": "approveEpisode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "El episodio, ya en APPROVED. La síntesis de audio arranca en segundo plano.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: el episodio no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "`INVALID_STATE_TRANSITION`: el episodio no está en un estado que admita la acción (api-contract.md §5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Aprobar (PENDING_REVIEW → APPROVED). Sin body.",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/episodes/{id}/actions/edit": {
      "post": {
        "operationId": "editEpisodeArgument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El argumento editado (`origin: HUMAN_EDITED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArgumentDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR`: body mal formado (falta un campo, tipo incorrecto o campos de más).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: el episodio no existe o el `argumentId` no es de un round de su debate. No se toca nada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "`INVALID_STATE_TRANSITION`: el episodio no está en un estado que admita la acción (api-contract.md §5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Editar el texto de un argumento (solo PENDING_REVIEW).",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/episodes/{id}/actions/reject": {
      "post": {
        "operationId": "rejectEpisode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "El episodio, ya en CANCELLED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: el episodio no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "`INVALID_STATE_TRANSITION`: el episodio no está en un estado que admita la acción (api-contract.md §5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Rechazar (PENDING_REVIEW o REQUIRES_HUMAN_REVIEW → CANCELLED). Sin body.",
        "tags": [
          "episodes"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Trend Debates API",
    "description": "Pipeline automatizado que convierte un trend en un debate verificado entre agentes de IA.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "atd_session",
        "description": "Cookie httpOnly que emite POST /auth/login (7 días)."
      }
    },
    "schemas": {
      "DebateLanguage": {
        "type": "string",
        "enum": [
          "ES",
          "EN",
          "PT"
        ]
      },
      "CreateEpisodeDto": {
        "type": "object",
        "properties": {
          "topic": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "language": {
            "default": "ES",
            "$ref": "#/components/schemas/DebateLanguage"
          }
        },
        "required": [
          "topic"
        ],
        "additionalProperties": false
      },
      "DebateLanguage_Output": {
        "type": "string",
        "enum": [
          "ES",
          "EN",
          "PT"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "EpisodeListItemDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "RESEARCHING",
              "READY_FOR_DEBATE",
              "DEBATING",
              "JUDGING",
              "PENDING_REVIEW",
              "APPROVED",
              "CANCELLED",
              "FAILED",
              "GENERATING_AUDIO",
              "READY_FOR_RENDER",
              "RENDERING",
              "COMPLETED",
              "REQUIRES_HUMAN_REVIEW"
            ]
          },
          "title": {
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/DebateLanguage_Output"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "title",
          "language",
          "createdAt",
          "publishedAt"
        ],
        "additionalProperties": false
      },
...
      "SignedAudioUrlDto_Output": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "additionalProperties": false
      },
...
    }
  }
}